+++ pyoxidizer/src/starlark/file_resource.rs
@@ -244,12 +244,7 @@ impl FileManifest {
             }
             "PythonResourceData" => {
                 let m = resource.downcast_apply(|m: &PythonResourceData| m.data.clone());
-                warn!(
-                    logger,
-                    "adding resource file {} to {}",
-                    m.full_name(),
-                    prefix
-                );
+                warn!(logger, "adding resource file {} to {}", m.full_name, prefix);
                 m.add_to_file_manifest(&mut self.manifest, &prefix)
                     .or_else(|e| {
                         Err(RuntimeError {
@@ -606,8 +601,9 @@ mod tests {
 
         let v = Value::new(PythonResourceData {
             data: ResourceData {
-                package: "foo.bar".to_string(),
-                name: "resource.txt".to_string(),
+                full_name: "foo/bar/resource.txt".to_string(),
+                leaf_package: "foo.bar".to_string(),
+                relative_name: "resource.txt".to_string(),
                 data: DataLocation::Memory(vec![]),
             },
         });